Refactored run algorithm procedure.
- Created utility function for getting matching data on the server before running algorithm
- Refactored app to not fetch and pass along matching data to run mutation on the client
- Created query procedure to get selectedAlgName on server
- Removed obsolete matching data query procedure
