+++ src/server/routers/institution/instance/matching.ts
@@ -17,108 +17,6 @@ import { instanceParamsSchema } from "@/lib/validations/params";
 import { adminProcedure, createTRPCRouter } from "@/server/trpc";
 
 export const matchingRouter = createTRPCRouter({
-  data: adminProcedure.input(z.object({ params: instanceParamsSchema })).query(
-    async ({
-      ctx,
-      input: {
-        params: { group, subGroup, instance },
-      },
-    }) => {
-      const studentData = await ctx.db.userInInstance.findMany({
-        where: {
-          allocationGroupId: group,
-          allocationSubGroupId: subGroup,
-          allocationInstanceId: instance,
-          role: Role.STUDENT,
-          submittedPreferences: true,
-        },
-        select: {
-          userId: true,
-          studentPreferences: {
-            where: { type: { equals: PreferenceType.PREFERENCE } },
-            select: { projectId: true, rank: true },
-            orderBy: { rank: "asc" },
-          },
-        },
-      });
-
-      const supervisorData = await ctx.db.supervisorInstanceDetails.findMany({
-        where: {
-          allocationGroupId: group,
-          allocationSubGroupId: subGroup,
-          allocationInstanceId: instance,
-        },
-        select: {
-          userId: true,
-          projectAllocationLowerBound: true,
-          projectAllocationTarget: true,
-          projectAllocationUpperBound: true,
-        },
-      });
-
-      const projectData = await ctx.db.project.findMany({
-        where: {
-          allocationGroupId: group,
-          allocationSubGroupId: subGroup,
-          allocationInstanceId: instance,
-        },
-        select: {
-          id: true,
-          supervisorId: true,
-          capacityLowerBound: true,
-          capacityUpperBound: true,
-        },
-      });
-
-      const students = studentData.map(({ userId, studentPreferences }) => ({
-        id: userId,
-        preferences: studentPreferences.map(({ projectId }) => projectId),
-      }));
-
-      const projects = projectData.map(
-        ({ id, supervisorId, capacityLowerBound, capacityUpperBound }) => ({
-          id,
-          lowerBound: capacityLowerBound,
-          upperBound: capacityUpperBound,
-          supervisorId,
-        }),
-      );
-
-      const supervisors = supervisorData.map(
-        ({
-          userId,
-          projectAllocationLowerBound,
-          projectAllocationTarget,
-          projectAllocationUpperBound,
-        }) => ({
-          id: userId,
-          lowerBound: projectAllocationLowerBound,
-          target: projectAllocationTarget,
-          upperBound: projectAllocationUpperBound,
-        }),
-      );
-
-      const data = { students, projects, supervisors };
-
-      const allocationInstance =
-        await ctx.db.allocationInstance.findFirstOrThrow({
-          where: {
-            allocationGroupId: group,
-            allocationSubGroupId: subGroup,
-            id: instance,
-          },
-          select: { selectedAlgName: true },
-        });
-
-      const selectedAlgName = allocationInstance?.selectedAlgName ?? undefined;
-
-      return {
-        matchingData: data,
-        selectedAlgName,
-      };
-    },
-  ),
-
   select: adminProcedure
     .input(
       z.object({
